+++ packages/core/src/lib/plugins/ParserPlugin.ts
@@ -34,6 +34,7 @@ export const ParserPlugin = createSlatePlugin({
             !pipeInsertDataQuery(editor, injectedPlugins, {
               data,
               dataTransfer,
+              mimeType,
             })
           ) {
             continue;
@@ -42,12 +43,14 @@ export const ParserPlugin = createSlatePlugin({
           data = pipeTransformData(editor, injectedPlugins, {
             data,
             dataTransfer,
+            mimeType,
           });
 
           let fragment = deserialize?.({
             ...getEditorPlugin(editor, plugin),
             data,
             dataTransfer,
+            mimeType,
           });
 
           if (!fragment?.length) continue;
@@ -56,6 +59,7 @@ export const ParserPlugin = createSlatePlugin({
             data,
             dataTransfer,
             fragment,
+            mimeType,
           });
 
           if (fragment.length === 0) continue;
@@ -64,6 +68,7 @@ export const ParserPlugin = createSlatePlugin({
             data,
             dataTransfer,
             fragment,
+            mimeType,
           });
 
           return true;