+++ packages/core/src/lib/static/plugins/getStaticPlugins.ts
@@ -1,0 +1,7 @@
+import { ViewPlugin } from './ViewPlugin';
+
+export const getStaticPlugins = () => {
+  const staticPlugins = [ViewPlugin];
+
+  return [...staticPlugins];
+};